+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/settings/ZigCoreProjectConfigurationProvider.kt
@@ -24,6 +24,7 @@ package com.falsepattern.zigbrains.project.settings
 
 import com.falsepattern.zigbrains.shared.SubConfigurable
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.project.ProjectManager
 
 class ZigCoreProjectConfigurationProvider: ZigProjectConfigurationProvider {
     override fun handleMainConfigChanged(project: Project) {
@@ -34,7 +35,7 @@ class ZigCoreProjectConfigurationProvider: ZigProjectConfigurationProvider {
     }
 
     override fun createNewProjectSettingsPanel(): ZigProjectConfigurationProvider.SettingsPanel {
-        return ZigProjectSettingsPanel(null)
+        return ZigProjectSettingsPanel(ProjectManager.getInstance().defaultProject)
     }
 
     override val priority: Int