+++ lsp/src/main/kotlin/com/falsepattern/zigbrains/lsp/settings/ZLSSettings.kt
@@ -22,22 +22,38 @@
 
 package com.falsepattern.zigbrains.lsp.settings
 
+import com.falsepattern.zigbrains.lsp.config.SemanticTokens
+import com.falsepattern.zigbrains.lsp.config.ZLSConfig
 import com.falsepattern.zigbrains.project.settings.ZigProjectConfigurationProvider
 import com.intellij.openapi.project.Project
 import org.jetbrains.annotations.NonNls
 
+@Suppress("PropertyName")
 data class ZLSSettings(
     var direnv: Boolean = false,
     var zlsPath: @NonNls String = "",
     var zlsConfigPath: @NonNls String = "",
-    var debug: Boolean = false,
-    var messageTrace: Boolean = false,
-    var buildOnSave: Boolean = false,
-    var buildOnSaveStep: @NonNls String = "install",
-    var globalVarDeclarations: Boolean = false,
-    var comptimeInterpreter: Boolean = false,
-    var inlayHints: Boolean = true,
-    var inlayHintsCompact: Boolean = true
+    val inlayHints: Boolean = true,
+    val enable_snippets: Boolean = true,
+    val enable_argument_placeholders: Boolean = true,
+    val completion_label_details: Boolean = true,
+    val enable_build_on_save: Boolean = false,
+    val build_on_save_args: List<String> = emptyList(),
+    val semantic_tokens: SemanticTokens = SemanticTokens.full,
+    val inlay_hints_show_variable_type_hints: Boolean = true,
+    val inlay_hints_show_struct_literal_field_type: Boolean = true,
+    val inlay_hints_show_parameter_name: Boolean = true,
+    val inlay_hints_show_builtin: Boolean = true,
+    val inlay_hints_exclude_single_argument: Boolean = true,
+    val inlay_hints_hide_redundant_param_names: Boolean = false,
+    val inlay_hints_hide_redundant_param_names_last_token: Boolean = false,
+    val warn_style: Boolean = false,
+    val highlight_global_var_declarations: Boolean = false,
+    val skip_std_references: Boolean = false,
+    val prefer_ast_check_as_child_process: Boolean = true,
+    val builtin_path: String? = null,
+    val build_runner_path: @NonNls String? = null,
+    val global_cache_path: @NonNls String? = null,
 ): ZigProjectConfigurationProvider.Settings {
     override fun apply(project: Project) {
         project.zlsSettings.loadState(this)